+++ src/modify/FMODAudioEngine.cpp
@@ -1,3 +1,7 @@
+/// @todo: Fix this code - on Windows its fine, but on Android it bugs out.
+/// I'm not going to work on this right now, but I'll leave it here for future reference. (@OmgRod)
+
+/*
 #include <Geode/Geode.hpp>
 #include <Geode/modify/FMODAudioEngine.hpp>
 
@@ -17,3 +21,4 @@ public:
         FMODAudioEngine::playMusic(name, shouldLoop, fadeInTime, channel);
     }
 };
+*/