+++ changelog.md
@@ -7,7 +7,7 @@
 - Bugfix: [Icon Ninja stuff (thanks undefined0 :D)](https://github.com/OmgRod/Geodify/issues/14)
 - Changed color offset to change [each individual wave](https://github.com/OmgRod/Geodify/issues/13)
 - Switched from External Hook API to Alpha's Geode Utils
-- Geodify layer previews
+- ~~Geodify layer previews~~ (delayed to v2.0.1 or v2.1.0)
 
 ```
 Setting keys were changed. What does this mean?